+++ pkg/resolver/cluster.go
@@ -4,6 +4,7 @@ type Cluster struct {
 	AdditionalTags              map[string]string
 	VPCsToAssociateToHostedZone []string
 	Name                        string
+	HostedZoneName              string
 	ControlPlaneEndpoint        string
 	IsDnsModePrivate            bool
 	IsVpcModePrivate            bool